+++ de.uni_koeln.spinfo.strings.core/src/de/uni_koeln/spinfo/strings/algo/suffixtrees/node/memory/SuffixNode.java
@@ -107,7 +107,7 @@ public abstract class SuffixNode implements Node{
 	}
 
 
-	public long getTextNumber() {
+	public int getTextNumber() {
 		return textNumber;
 	}
